+++ packages/react-scripts/config/webpack.config.dev.js
@@ -178,7 +178,11 @@ module.exports = {
       //   include: paths.appSrc,
       // },
       {
-        test: /(^|\/)lodash(\/|\.|$)/,
+        test: new RegExp(
+          `(^|/)(${(bpkReactScriptsConfig.amdExcludes || [])
+            .concat('lodash')
+            .join('|')})(/|.|$)`
+        ),
         parser: {
           amd: false,
         },