+++ packages/react-scripts/config/webpack.config.prod.js
@@ -188,7 +188,11 @@ module.exports = {
       //   include: paths.appSrc,
       // },
       {
-        test: /(^|\/)lodash(\/|\.|$)/,
+        test: new RegExp(
+          `(^|/)(${(bpkReactScriptsConfig.amdExcludes || [])
+            .concat('lodash')
+            .join('|')})(/|.|$)`
+        ),
         parser: {
           amd: false,
         },